+++ crates/snxcore/src/tunnel.rs
@@ -2,7 +2,6 @@ use std::{sync::Arc, time::Duration};
 
 use async_trait::async_trait;
 use bytes::Bytes;
-use ipnet::Ipv4Net;
 use tokio::sync::mpsc;
 
 use crate::model::{
@@ -30,7 +29,7 @@ pub enum TunnelEvent {
     Disconnected,
     RekeyCheck,
     RemoteControlData(Bytes),
-    Rekeyed(Ipv4Net),
+    Rekeyed(IPsecSession),
     Rtt(Duration),
 }
 
@@ -60,6 +59,7 @@ pub trait TunnelConnector {
     ) -> anyhow::Result<Box<dyn VpnTunnel + Send>>;
     async fn terminate_tunnel(&mut self, signout: bool) -> anyhow::Result<()>;
     async fn handle_tunnel_event(&mut self, event: TunnelEvent) -> anyhow::Result<()>;
+    async fn rekey(&mut self) -> anyhow::Result<()>;
 }
 
 #[async_trait]